engine/Oldrequirements.txt
dill==0.3.3
skater==1.0.4
minio==5.0.7
mlflow==1.8.0
imbalanced-learn==0.6.2
plotly==4.8.1

# maybe you will need to.. sudo xcodebuild -license
matplotlib==3.2.1
Pillow
# Requirements automatically generated by pigar.
# https://github.com/damnever/pigar

# pipeline_sample_code/pipeline_test.py: 20,21,22,24
Flask == 2.0.0

# pipeline_sample_code/pipeline_test.py: 23
Flask_Cors == 3.0.8

# pipeline_sample_code/dagster_types/model_type.py: 25
boto3 == 1.9.253

# pipeline_sample_code/dagster_resources/s3_resource.py: 3
botocore == 1.12.253

# pipeline_sample_code/dagster_repo/repos.py: 1
# pipeline_sample_code/dagster_resources/nothing.py: 1
# pipeline_sample_code/dagster_resources/s3_resource.py: 1
# pipeline_sample_code/dagster_types/dataframe_type.py: 19
# pipeline_sample_code/dagster_types/model_type.py: 19
# pipeline_sample_code/main.py: 14
# pipeline_sample_code/pipeline_test.py: 14
# pipeline_sample_code/tasks/solids.py: 17,20
# pipeline_sample_code/tests/solids_test.py: 1
# pipeline_sample_code/tests/test_dagster.py: 1
dagster == 0.7.6

dagster_postgres == 0.7.6

dagster_cron == 0.7.6

# pipeline_sample_code/dagster_types/dataframe_type.py: 57
# pipeline_sample_code/dagster_types/model_type.py: 53
# pipeline_sample_code/main.py: 13,19
# pipeline_sample_code/pipeline_test.py: 13,19
# pipeline_sample_code/tasks/solids.py: 30,31
dagster_aws == 0.7.6

# pipeline_sample_code/dagster_types/dataframe_type.py: 22
# pipeline_sample_code/dagster_types/model_type.py: 22
# pipeline_sample_code/main.py: 12
# pipeline_sample_code/pipeline_test.py: 12
# pipeline_sample_code/tasks/solids.py: 19
# pipeline_sample_code/tests/solids_test.py: 10
# pipeline_sample_code/tests/test_dagster.py: 10
dagster_pyspark == 0.7.6
dagster_spark == 0.7.6

# pipeline_sample_code/dagster_types/dataframe_type.py: 30
# pipeline_sample_code/dagster_types/model_type.py: 26
# pipeline_sample_code/pipeline_test.py: 25
joblib~=1.1.0

# pipeline_sample_code/main.py: 1
# pipeline_sample_code/pipeline_test.py: 1
# pipeline_sample_code/tasks/solids.py: 1,10,16,25,198
# pipeline_sample_code/tests/tasks_pipeline_test.py: 7
numpy~=1.19.5

# pipeline_sample_code/dagster_types/dataframe_type.py: 26
# pipeline_sample_code/main.py: 17
# pipeline_sample_code/pipeline_test.py: 17
# pipeline_sample_code/tasks/solids.py: 11,14,26,36
# pandas == 1.0.3
pandas=1.3.4

# pipeline_sample_code/dagster_types/dataframe_type.py: 24,28,29
# pipeline_sample_code/tasks/solids.py: 8,9,12,23,24,27
# pyarrow == 0.16.0
pyarrow == 5.0.0
parquet == 1.3.1



# pipeline_sample_code/dagster_types/dataframe_type.py: 3,4
# pipeline_sample_code/dagster_types/model_type.py: 3,4,23
# pipeline_sample_code/tasks/heavy.py: 8
# pipeline_sample_code/tests/solids_test.py: 6,7,9
# pipeline_sample_code/tests/tasks_pipeline_test.py: 13
# pipeline_sample_code/tests/test_dagster.py: 6,7,9
# pipeline_sample_code/tests/test_minio.py: 1,2
# pyspark == 2.4.4
pyspark == 3.1.2

# pipeline_sample_code/dagster_resources/s3_resource.py: 2
# pipeline_sample_code/dagster_types/dataframe_type.py: 25,27
# pipeline_sample_code/dagster_types/model_type.py: 27
s3fs == 0.4.2

# pipeline_sample_code/main.py: 16
# pipeline_sample_code/pipeline_test.py: 16
# pipeline_sample_code/tasks/light.py: 1,2,3,4,5,6,7
scikit_learn ==0.23.2

# pipeline_sample_code/main.py: 10
# pipeline_sample_code/pipeline_test.py: 10
six == 1.14.0
psycopg2-binary

scikit-learn~=0.23.2
xgboost~=1.6.1
pytest~=5.3.5
requests~=2.22.0
torch~=1.12.0
pycaret~=2.3.10